Begin sketching the injection of Serlain into captured sites' code
To e.g. detect the user clicking on a link within the captured site.

This can eventually allow e.g. the mock browser's address bar to keep
synced to the browsing flow of the iframe; but also to prevent e.g.
target="_top" on links (which would navigate the main Serlain page
instead of the iframe).